Created static formatDate function in App.js
