+++ client/src/components/Profile.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import MeetAttending from './MeetAttending';
 import './Profile.css';
+import App from './App';
 
 class Profile extends Component {
   render() {
@@ -25,7 +26,7 @@ class Profile extends Component {
           <div className="data-pair">
             <h4>Date of Birth:</h4>
             <div className="profile-data">
-              <p>{formatDate(this.props.user.birth)}</p>
+              <p>{App.formatDate(this.props.user.birth)}</p>
             </div>
           </div>
           <div className="data-pair">
@@ -57,13 +58,5 @@ class Profile extends Component {
   }
 }
 
-// format the date for better readability
-function formatDate(dateString) {
-  if (dateString){
-  var shortened = dateString.slice(0, 10);
-  var dateParts= shortened.split("-");
-  return Number(dateParts[2]) + 1 + "." + dateParts[1] + "." + dateParts[0];
-  } else return null;
-};
 
 export default Profile;
