+++ client/src/components/ProfileForm.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import axios from 'axios';
 import DatePicker from 'react-datepicker';
 import moment from 'moment';
+import App from './App';
 
 import 'react-datepicker/dist/react-datepicker.css';
 import './Profile.css';
